adds WheatParser to hub
