+++ mantisshrimp/hub/parsers/__init__.py
@@ -1,0 +1,1 @@
+from .wheat_parser import *
