+++ mantisshrimp/hub/parsers/wheat_parser.py
@@ -1,0 +1,35 @@
+__all__ = ["WheatParser"]
+
+from mantisshrimp.imports import *
+from mantisshrimp import *
+
+
+class WheatParser(DefaultImageInfoParser, FasterRCNNParser):
+    def __init__(self, df, source):
+        self.df = df
+        self.source = source
+        self.imageid_map = IDMap()
+
+    def __iter__(self):
+        yield from self.df.itertuples()
+
+    def __len__(self):
+        return len(self.df)
+
+    def imageid(self, o) -> int:
+        return self.imageid_map[o.image_id]
+
+    def filepath(self, o) -> Union[str, Path]:
+        return self.source / f"{o.image_id}.jpg"
+
+    def height(self, o) -> int:
+        return o.height
+
+    def width(self, o) -> int:
+        return o.width
+
+    def label(self, o) -> List[int]:
+        return [1]
+
+    def bbox(self, o) -> List[BBox]:
+        return [BBox.from_xywh(*np.fromstring(o.bbox[1:-1], sep=","))]
